refactor: move states and functions for course details to separate hook useCourseActions
