+++ src/functions/index.ts
@@ -3,3 +3,4 @@ export { capitalizeFirstLetter } from './capitalizeFirstLetter';
 export { groupCoursesByCategory } from './course';
 export { handleAuthError } from './errorHandlers';
 export { getCSSVariableValue } from './getCSSVariableValue';
+export { getCategoryLabel, getSubcategoryLabel } from './getLabels';
