+++ src/hooks/index.ts
@@ -1,5 +1,6 @@
 export { useAsyncInitialize } from './useAsyncInitialize';
-export { useContract } from './useContract';
+export { useCourseActions } from './useCourseActions';
+export { useCourseContract } from './useCourseContract';
 export { useCourseForm } from './useCourseForm';
 export { useCoursePartForm } from './useCoursePartForm';
 export { useTonClient } from './useTonClient';
