+++ src/hooks/useCourseContract.tsx
@@ -2,8 +2,9 @@ import { OpenedContract, address, fromNano, toNano } from '@ton/core';
 import { useCallback, useEffect, useState } from 'react';
 import TonWeb from 'tonweb';
 import { Course, NewCourse } from '../ton/wrappers/Course';
-import { DeployType } from '../types';
+import { ICourse, RoleType } from '../types';
 import { useAsyncInitialize } from './useAsyncInitialize';
+import { useCourseActions } from './useCourseActions';
 import { useTonClient } from './useTonClient';
 import { useTonConnect } from './useTonConnect';
 
@@ -16,16 +17,17 @@ const tonweb = new TonWeb(
   )
 );
 
-export function useContract(courseId: string, coursePrice: number) {
+export function useCourseContract(course: ICourse, role: RoleType) {
+  const courseId = course.id;
   const { client } = useTonClient();
   const { sender } = useTonConnect();
 
   const [balance, setBalance] = useState<string>('0');
-  const [modalOpen, setModalOpen] = useState<boolean>(false);
-  const [deployType, setDeployType] = useState<DeployType | null>(null);
   const [errorContract, setErrorContract] = useState<string>('');
 
-  const coursePriceInNano = toNano(coursePrice.toString());
+  const { handleAddPointsForCreating } = useCourseActions(course, role);
+
+  const coursePriceInNano = toNano(course.price.toString());
 
   const courseDefaultContract = useAsyncInitialize(async () => {
     if (!client) return;
@@ -68,26 +70,22 @@ export function useContract(courseId: string, coursePrice: number) {
     setBalance(fromNano(contractBalance));
   }, [getContractBalance]);
 
-  const checkAndUpdateBalance = useCallback(
-    async (type: DeployType) => {
-      const initialBalance = await getContractBalance();
-      setBalance(fromNano(initialBalance));
+  const checkAndUpdateBalance = useCallback(async () => {
+    const initialBalance = await getContractBalance();
+    setBalance(fromNano(initialBalance));
 
-      const intervalId = setInterval(async () => {
-        const currentBalance = await getContractBalance();
-        if (currentBalance !== initialBalance) {
-          setBalance(fromNano(currentBalance));
-          setModalOpen(true);
-          setDeployType(type);
-          clearInterval(intervalId);
-        }
-      }, 5000); // 5 sec
+    const intervalId = setInterval(async () => {
+      const currentBalance = await getContractBalance();
+      if (currentBalance !== initialBalance) {
+        setBalance(fromNano(currentBalance));
+        await handleAddPointsForCreating();
+        clearInterval(intervalId);
+      }
+    }, 5000); // 5 sec
 
-      // Clear the interval when unmounting a component or changing dependencies
-      return () => clearInterval(intervalId);
-    },
-    [getContractBalance]
-  );
+    // Clear the interval when unmounting a component or changing dependencies
+    return () => clearInterval(intervalId);
+  }, [getContractBalance, handleAddPointsForCreating]);
 
   useEffect(() => {
     updateBalance();
@@ -96,9 +94,6 @@ export function useContract(courseId: string, coursePrice: number) {
   return {
     balance,
     errorContract,
-    deployType,
-    modalOpen,
-    setModalOpen,
     createCourse: async () => {
       const message: NewCourse = {
         $$type: 'NewCourse',
@@ -113,7 +108,7 @@ export function useContract(courseId: string, coursePrice: number) {
           },
           message
         );
-        await checkAndUpdateBalance('create');
+        await checkAndUpdateBalance();
       } catch (error: any) {
         setErrorContract(
           `Transaction failed or was rejected. ${error?.message}`
@@ -130,7 +125,7 @@ export function useContract(courseId: string, coursePrice: number) {
           },
           'New purchase'
         );
-        await checkAndUpdateBalance('purchase');
+        await checkAndUpdateBalance();
       } catch (error: any) {
         setErrorContract(
           `Transaction failed or was rejected. ${error?.message}`
